backend/core/requirements.txt
python-decouple==3.8
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-environ==0.11.2
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
Pillow
